--- packages/gamut/src/ProvidedForm/index.tsx
@@ -1,2 +1,0 @@
-export * from './FormField';
-export * from './FormWrapper';
